% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESeq2MultiAssayDS.R
\name{DESeq2MultiAssayDS}
\alias{DESeq2MultiAssayDS}
\title{Differential expression analysis on a MultiAssayExperiment; server-side}
\usage{
DESeq2MultiAssayDS(
  vars,
  MultiAssay,
  SE,
  test,
  fitType,
  sfType,
  reduced,
  contrast
)
}
\arguments{
\item{vars}{variables for the model;}

\item{MultiAssay}{a MultiAssayExperiment object;}

\item{SE}{a character string containing the name of a SummarizedExperiment object;}

\item{test}{"Wald" or "LRT";}

\item{fitType}{"parametric", "local", or "mean";}

\item{sfType}{"ratio", "poscounts", or "iterate";}

\item{reduced}{test="LRT";}

\item{contrast}{model matrix contrast.}
}
\description{
Given a MultiAssayExperiment object and the name of one of its SummarizedExperiment,
the function performs a Differential gene expression analysis using \code{DESeq2} package.
}
