% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimentsDS.R
\name{experimentsDS}
\alias{experimentsDS}
\title{Extract Experiments from a MultiAssayExperiment; server-side}
\usage{
experimentsDS(Multiassay)
}
\arguments{
\item{Multiassay}{A non NULL MultiAssayExperiment object.}
}
\value{
An ExperimentList object
}
\description{
Given a MultiAssayExperiment object, the function returns the corresponding ExperimentList
containing the experiments data.
}
