% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterCompleteCasesDS.R
\name{filterCompleteCasesDS}
\alias{filterCompleteCasesDS}
\title{Filter a MultiAssayExperiment; server-side}
\usage{
filterCompleteCasesDS(MultiAssay)
}
\arguments{
\item{MultiAssay}{A MultiAssayExperiment object.}
}
\value{
A filtered MultiAssayExperiment object, containing only complete cases.
}
\description{
Given a MultiAssayExperiment object, the function retains only the columns present across all experiments
and returns a filtered MultiAssayExperiment object.
Before filtering the object, the function checks the input is valid.
}
