% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiAssay_check_missing_dataDS.R
\name{MultiAssay_check_missing_dataDS}
\alias{MultiAssay_check_missing_dataDS}
\title{Remove biological units with missing values from a MultiAssayExperiment object}
\usage{
MultiAssay_check_missing_dataDS(MultiAssay, coldatafilt = FALSE)
}
\arguments{
\item{MultiAssay}{a MultiAssayExperiment object;}

\item{coldatafilt}{logical whether to filter also colData (default is FALSE).}
}
\value{
a MultiAssayExperiment object without biological units containing NA values.
}
\description{
For each assay of a given MultiAssayExperiment object, it removes biological units
containing at least one NA.
}
