% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiAssay_preprocess_dataDS.R
\name{MultiAssay_preprocess_dataDS}
\alias{MultiAssay_preprocess_dataDS}
\title{Remove Features with missing values exceeding a given threshold}
\usage{
MultiAssay_preprocess_dataDS(MultiAssay, vect_thr, coldata_thr = 0)
}
\arguments{
\item{MultiAssay}{A MultiAssayExperiment object;}

\item{vect_thr}{Vector of thresholds;}

\item{coldata_thr}{Integer indicating the threshold to be applied to colData (default 0).}
}
\value{
A filtered MultiAssayExperiment object
}
\description{
For each assay of a given MultiAssayExperiment object, it removes features containing NA over a given threshold.
Thresholds vector can be of three types:
\itemize{
\item single value between 0 and 100: the same value will be applied to all the assays;
\item vector of numerical values (from 0 to 100) having the same length of MultiAssayExperiment: each value will be applied to the corresponding assay;
\item vector containing the names of the assay with the corresponding threshold to be applied (from 0 to 100). (e.g. 'RNASeqdataa'=50)
}
}
