% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiAssayDS.R
\name{MultiAssayDS}
\alias{MultiAssayDS}
\title{Create a MultiAssayExperiment Object}
\usage{
MultiAssayDS(
  exp,
  df = NULL,
  df_col = NULL,
  colD = NULL,
  levels_col = NULL,
  level = NULL,
  sampleMap = NULL,
  metadata = list(),
  drops = list()
)
}
\arguments{
\item{exp}{A list or ExperimentList of experiments;}

\item{df}{Optional data frame to be processed into colData or new assays;}

\item{df_col}{character, the column name in \code{df} containing biological units identifiers;}

\item{colD}{dataframe with characteristics of biological units;}

\item{levels_col}{character, the column in \code{df} used to split data into multiple assays;}

\item{level}{optional Vector of specific levels to be extracted from \code{levels_col};}

\item{sampleMap}{A data frame of assay names, sample identifiers, and colname samples;}

\item{metadata}{Optional list of metadata;}

\item{drops}{Optional list for dropping elements.}
}
\value{
A MultiAssayExperiment object
}
\description{
Given a list of experiments, optional colData, sampleMap and metadata, the function constructs a MultiAssayExperiment object.
It can also integrate data from an input data frame, extracting colData and splitting the remaining data into multiple assays based on factor levels.
}
