% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_dfDS.R
\name{split_dfDS}
\alias{split_dfDS}
\title{Split a Data frame into variable and constant columns; server-side}
\usage{
split_dfDS(df, column)
}
\arguments{
\item{df}{A data frame to be processed}

\item{column}{A character string indicating the name of the grouping column}
}
\value{
A list of two data frames
}
\description{
Given a data frame and a reference grouping column, the function aggregates all rows of the same group,
identifying, for each group, variable and constant columns.
The function returns a list containing two data frames:
\itemize{
\item filtered_df, i.e. the original data frame with constant columns removed;
\item filtered_coldata, i.e. a new data frame containing only the constants columns.
}
}
\keyword{internal}
