% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statCompleteCasesDS.R
\name{statCompleteCasesDS}
\alias{statCompleteCasesDS}
\title{Statistics of Complete cases in a MultiAssayExperiment; server-side}
\usage{
statCompleteCasesDS(MultiAssay)
}
\arguments{
\item{MultiAssay}{a MultiAssayExperiment object.}
}
\value{
A table with the number of incomplete and complete cases
}
\description{
Given a MultiAssayExperiment object, the function returns a frequency table containing the number of
complete cases across all experiments.
}
